requirements.txt
aiofiles==24.1.0
aiohttp==3.8.6
azure_storage==0.37.0
boto3==1.34.14
chainlit==1.3.1
dill==0.3.8
elasticsearch==8.12.1
fastapi==0.115.4
jieba==0.42.1
langdetect==1.0.9
literalai==0.0.601
mypy_boto3_dynamodb==1.35.24
nltk==3.8.1
openai==0.28.0
pandas==2.2.3
psycopg2_binary==2.9.9
PyMySQL==1.1.1
Requests==2.32.3
rouge==1.0.1
sentence_transformers==2.6.1
SQLAlchemy==1.4.51
sqlparse==0.5.1
tabulate==0.9.0
uvicorn==0.32.0
